+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Electron 17.2.0 (was 17.0.1)
 - Chromium 98.0.4758.109 (was 98.0.4758.82)
 - Permissions for serial and other human interface devices (permissionserial, permissionhid, both blocked by default, same as before)
+- Attempt to delete blocked navigator properties from navigator object completely instead of returning undefined
 
 ## [7.1.2](https://github.com/Jelmerro/Vieb/compare/7.1.1...7.1.2) - 2022-02-20